+++ lib/article_json/export/amp/elements/embed.rb
@@ -12,10 +12,18 @@ module ArticleJSON
             end
           end
 
-          def amp_library
-            return @amp_library if defined? @amp_library
-            type_specific_node # the execution will determine needed amp library
-            @amp_library
+          # Returns the custom element tags required for this embedded element
+          #
+          # @return [Array[Symbol]]
+          def custom_element_tags
+            case @element.embed_type.to_sym
+            when :youtube_video then %i(amp-youtube)
+            when :vimeo_video then %i(amp-vimeo)
+            when :facebook_video then %i(amp-facebook)
+            when :tweet then %i(amp-twitter)
+            when :slideshare then %i(amp-iframe)
+            else []
+            end
           end
 
           private
@@ -41,52 +49,28 @@ module ArticleJSON
             end
           end
 
-          def youtube_library
-            '<script async custom-element="amp-youtube" ' \
-            'src="https://cdn.ampproject.org/v0/amp-youtube-0.1.js"></script>'
-          end
-
           def youtube_node
-            @amp_library = youtube_library
             create_element('amp-youtube',
                            'data-videoid' => @element.embed_id,
                            width: default_width,
                            height: default_height)
           end
 
-          def vimeo_library
-            '<script async custom-element="amp-vimeo"' \
-            'src="https://cdn.ampproject.org/v0/amp-vimeo-0.1.js"></script>'
-          end
-
           def vimeo_node
-            @amp_library = vimeo_library
             create_element('amp-vimeo',
                            'data-videoid' => @element.embed_id,
                            width: default_width,
                            height: default_height)
           end
 
-          def tweet_library
-            '<script async custom-element="amp-twitter" ' \
-            'src="https://cdn.ampproject.org/v0/amp-twitter-0.1.js"></script>'
-          end
-
           def tweet_node
-            @amp_library = tweet_library
             create_element('amp-twitter',
                            'data-tweetid' => @element.embed_id,
                            width: default_width,
                            height: default_height)
           end
 
-          def facebook_library
-            '<script async custom-element="amp-facebook" ' \
-            'src="https://cdn.ampproject.org/v0/amp-facebook-0.1.js"></script>'
-          end
-
           def facebook_node
-            @amp_library = facebook_library
             url = "#{@element.oembed_data[:author_url]}videos/#{@element.embed_id}"
             create_element('amp-facebook',
                            'data-embedded-as' => 'video',
@@ -95,13 +79,7 @@ module ArticleJSON
                            height: default_height)
           end
 
-          def iframe_library
-            '<script async custom-element="amp-iframe"' \
-            'src="https://cdn.ampproject.org/v0/amp-iframe-0.1.js"></script>'
-          end
-
           def slideshare_node
-            @amp_library = iframe_library
             node = Nokogiri::HTML(@element.oembed_data[:html]).xpath('//iframe')
             create_element('amp-iframe',
                            src: node.attribute('src').value,